token for activation completed
